removed redundant page
