+++ 19-AdvancedRouting/frontend/src/App.js
@@ -2,8 +2,8 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import RootLayout from "./pages/root";
 import ErrorPage from "./components/Error";
 import HomePage from "./pages/Main";
-import EventsPage from "./pages/Events";
 import EventItem from "./components/EventItem";
+import EventsList from "./components/EventsList";
 
 const router = createBrowserRouter([
   {
@@ -12,7 +12,7 @@ const router = createBrowserRouter([
     errorElement: <ErrorPage />,
     children: [
       { index: true, element: <HomePage /> },
-      { path: "/events", element: <EventsPage /> },
+      { path: "/events", element: <EventsList /> },
       { path: "/events/:eventId", element: <EventItem /> },
     ],
   },
